+++ app/study/views.py
@@ -7,54 +7,11 @@ from flask import render_template, jsonify, redirect, request, url_for, abort, f
 from flask_login import login_required, current_user
 from . import study
 from .. import db
-from ..models import Device
 from ..models import LessonType, Lesson, Video
-from ..models import VideoCollection, Collection
 from ..models import Punch
 from ..decorators import permission_required
 from ..utils import get_mac_address_from_ip
 from ..utils2 import add_user_log
-
-
-@study.route('/collection')
-def collection():
-    '''study.collection()'''
-    mac_address = get_mac_address_from_ip(ip_address=request.headers.get('X-Forwarded-For', request.remote_addr))
-    if mac_address is None:
-        flash('无法获取设备信息', category='error')
-        return redirect(url_for('auth.login'))
-    device = Device.query.filter_by(mac_address=mac_address).first()
-    if device is None:
-        flash('设备未授权（MAC地址：{}）'.format(mac_address), category='error')
-        return redirect(url_for('auth.login'))
-    if not device.restricted_permit:
-        flash('该设备无法访问受限资源', category='error')
-        return redirect(url_for('auth.login'))
-    collection_name = request.args.get('name')
-    if collection_name is None:
-        abort(404)
-    collection = Collection.query.filter_by(name=collection_name).first()
-    if collection is None:
-        abort(404)
-    query = Video.query\
-        .join(VideoCollection, VideoCollection.video_id == Video.id)\
-        .join(Collection, Collection.id == VideoCollection.collection_id)\
-        .filter(Collection.name == collection.name)\
-        .order_by(Video.id.asc())
-    video_id = request.args.get('video_id')
-    if video_id is not None:
-        video = Video.query.get_or_404(video_id)
-        if not collection.has_video(video=video):
-            abort(403)
-    else:
-        video = query.first()
-    videos = query.all()
-    return minify(render_template(
-        'study/collection.html',
-        collection=collection,
-        video=video,
-        videos=videos
-    ))
 
 
 @study.route('/vb')
@@ -62,14 +19,14 @@ def collection():
 @permission_required('研修VB')
 def vb():
     '''study.vb()'''
-    header = 'VB研修'
+    lesson_type = 'VB'
     lessons = Lesson.query\
         .join(LessonType, LessonType.id == Lesson.type_id)\
-        .filter(LessonType.name == 'VB')\
+        .filter(LessonType.name == lesson_type)\
         .order_by(Lesson.id.asc())
     return minify(render_template(
         'study/lesson.html',
-        header=header,
+        header=lesson_type,
         lessons=lessons
     ))
 
@@ -79,14 +36,63 @@ def vb():
 @permission_required('研修Y-GRE')
 def y_gre():
     '''study.y_gre()'''
-    header = 'Y-GRE研修'
+    lesson_type = 'Y-GRE'
     lessons = Lesson.query\
         .join(LessonType, LessonType.id == Lesson.type_id)\
-        .filter(LessonType.name == 'Y-GRE')\
+        .filter(LessonType.name == lesson_type)\
         .order_by(Lesson.id.asc())
     return minify(render_template(
         'study/lesson.html',
-        header=header,
+        header=lesson_type,
+        lessons=lessons
+    ))
+
+
+@study.route('/y-gre-aw')
+@login_required
+@permission_required('研修Y-GRE')
+def y_gre_aw():
+    '''study.y_gre_aw()'''
+    lesson_type = 'Y-GRE AW'
+    lessons = Lesson.query\
+        .join(LessonType, LessonType.id == Lesson.type_id)\
+        .filter(LessonType.name == lesson_type)\
+        .order_by(Lesson.id.asc())
+    return minify(render_template(
+        'study/lesson.html',
+        header=lesson_type,
+        lessons=lessons
+    ))
+
+
+@study.route('/test-review')
+@login_required
+@permission_required('研修Y-GRE')
+def test_review():
+    '''study.test_review()'''
+    lesson_type = '考试讲解'
+    lessons = Lesson.query\
+        .join(LessonType, LessonType.id == Lesson.type_id)\
+        .filter(LessonType.name == lesson_type)\
+        .order_by(Lesson.id.asc())
+    return minify(render_template(
+        'study/lesson.html',
+        header=lesson_type,
+        lessons=lessons
+    ))
+
+
+@study.route('/demo')
+def demo():
+    '''study.demo()'''
+    lesson_type = '体验课程'
+    lessons = Lesson.query\
+        .join(LessonType, LessonType.id == Lesson.type_id)\
+        .filter(LessonType.name == lesson_type)\
+        .order_by(Lesson.id.asc())
+    return minify(render_template(
+        'study/lesson.html',
+        header=lesson_type,
         lessons=lessons
     ))
 
@@ -97,10 +103,6 @@ def y_gre():
 def video(id):
     '''study.video(id)'''
     video = Video.query.get_or_404(id)
-    if video.restricted:
-        abort(403)
-    if not current_user.can('研修{}'.format(video.lesson.type.name)):
-        abort(403)
     if not current_user.can_study(lesson=video.lesson):
         flash('请先完成本课程的前序内容！', category='warning')
         return redirect(url_for('study.{}'.format(video.lesson.type.name_snake_case)))
@@ -108,7 +110,7 @@ def video(id):
         add_user_log(user=current_user._get_current_object(), event='视频研修：{}'.format(video.name), category='study')
         db.session.commit()
     return minify(render_template(
-        'study/video.html',
+        'study/video_demo.html',
         video=video
     ))
 
@@ -119,8 +121,6 @@ def video(id):
 def punch(id):
     '''study.punch(id)'''
     video = Video.query.get_or_404(id)
-    if video.restricted:
-        abort(403)
     if not current_user.can('研修{}'.format(video.lesson.type.name)) or not current_user.can_study(lesson=video.lesson):
         abort(403)
     if request.json is None: